Data log (GroundState object upload) to Parse. MT9D111 firmware state check for capture mode included in capture_cmd_list

+++ Proto1/Project/ULP_SEC_CAM/Application/Include/app_common.h
@@ -11,7 +11,7 @@
 #include "app.h"
 #include "math.h"
 
-unsigned short intToASCII(short cNum, char *cString);
+unsigned short intToASCII(long long cNum, char *cString);
 void LED_Blink(uint8_t ucHowManyTimes, float_t fSecsForEachCycle);
 void LED_Off();
 void LED_On();
